+++ db.json
@@ -1,24 +1,6 @@
 {
   "tasks": [
     {
-      "id": 3,
-      "name": "Crissie Summerill123",
-      "username": "csummerill2",
-      "title": "VP Product Management",
-      "date": "2022-01-26T21:46:27.765Z",
-      "value": 464.54
-    },
-    {
-      "id": 4,
-      "name": "Letitia Crolly",
-      "username": "lcrolly3",
-      "title": "Web Developer I",
-      "value": 183.58,
-      "date": "2021-07-10T20:39:48Z",
-      "image": "https://robohash.org/estveniamet.png?size=150x150&set=set1",
-      "isPayed": false
-    },
-    {
       "id": 5,
       "name": "Anthea Pundy",
       "username": "apundy4",
@@ -1667,6 +1649,14 @@
       "date": "2021-05-05T10:22:13Z",
       "image": "https://robohash.org/illumexpeditadeleniti.png?size=150x150&set=set1",
       "isPayed": true
+    },
+    {
+      "id": 173,
+      "name": "Adriana Lima Shikasho",
+      "username": "adrianashikasho2022",
+      "title": "front-end dev",
+      "date": "2022-01-26T23:14:26.857Z",
+      "value": "12300"
     }
   ],
   "account": [
