+++ src/app/service/user.service.ts
@@ -8,14 +8,10 @@ export class UserService {
   constructor(private http: HttpClient) {}
 
   getAll() {
-    return this.http.get<User[]>(`${"http://localhost:3000/tasks"}/account`)
+    return this.http.get<User[]>("http://localhost:3000/account")
   }
 
   register(user: User) {
-    return this.http.post(`${"http://localhost:3000/tasks"}/users/register`, user)
-  }
-
-  delete(id: number) {
-    return this.http.delete(`${"http://localhost:3000/tasks"}/users/${id}`)
+    return this.http.post("http://localhost:3000/account", user)
   }
 }
