+++ src/app/shared/error-msg/error-msg.component.html
@@ -1,4 +1,4 @@
-<div *ngIf="errorMessage != null">
+<div *ngIf="errorMessage !== null">
   <span class="glyphicon glyphicon-remove form-control-feedback"></span>
   <div class="alert alert-danger errorDiv" role="alert">
     {{ errorMessage }}
