--- src/app/shared/search/search.component.html
@@ -1,17 +1,0 @@
-<form class="search">
-  <div class="form-group__flex-columm">
-    <div class="input-relative-container">
-      <input
-        #input
-        (keyup)="search(input.value)"
-        class="form-control"
-        type="text"
-        placeholder="Pesquisar por usuário"
-      />
-
-      <div class="search-icon">
-        <img src="assets/icons/search.svg" alt="Buscar Tarefa" />
-      </div>
-    </div>
-  </div>
-</form>
