--- src/app/shared/search/search.component.scss
@@ -1,17 +1,0 @@
-:host {
-  form {
-    justify-content: space-between;
-
-    input {
-      width: 25rem;
-      height: 3.5rem;
-    }
-
-    .search-icon {
-      position: absolute;
-      top: 1.2rem;
-      right: 1rem;
-      cursor: pointer;
-    }
-  }
-}
