--- src/app/shared/search/task-search.component.spec.ts
@@ -1,24 +1,0 @@
-import { ComponentFixture, TestBed } from "@angular/core/testing";
-
-import { SearchComponent } from "./search.component";
-
-describe("SearchComponent", () => {
-  let component: SearchComponent;
-  let fixture: ComponentFixture<SearchComponent>;
-
-  beforeEach(async () => {
-    await TestBed.configureTestingModule({
-      declarations: [SearchComponent],
-    }).compileComponents();
-  });
-
-  beforeEach(() => {
-    fixture = TestBed.createComponent(SearchComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it("should create", () => {
-    expect(component).toBeTruthy();
-  });
-});
