+++ src/app/shared/toast/toast.component.scss
@@ -8,4 +8,5 @@
   top: 1rem;
   font-size: 1.6rem;
   transition: all 200ms;
+  z-index: 99999;
 }
